pyproject.toml
[project]
name = "ai-storyboard"
version = "0.1.0"
description = "AI-Driven Storyboard Generation: Extending Text-to-Shot Diffusion for Multi-Shot Visual Pre-production with Camera Motion Trajectories"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }

dependencies = [
    # Deep Learning
    "torch>=2.0.0",
    "torchvision>=0.15.0",

    # Human Body Model
    "smplx>=0.1.28",
    "numpy>=1.24.0",

    # Text Processing
    "transformers>=4.30.0",
    "sentence-transformers>=2.2.0",
    "openai>=1.0.0",

    # Visualization
    "matplotlib>=3.7.0",
    "Pillow>=9.5.0",
    "open3d>=0.17.0",
    "trimesh>=3.22.0",

    # Experiment Tracking
    "tensorboard>=2.13.0",
    "wandb>=0.15.0",

    # Data Processing
    "scipy>=1.10.0",
    "tqdm>=4.65.0",
    "PyYAML>=6.0",
    "einops>=0.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.1.0",
    "ipykernel>=6.0",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
